search_agent/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic>=2.0.0
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-openai
chromadb
pypdf
python-dotenv
rank_bm25
redis
pymilvus
tenacity
duckduckgo-search
docx2txt